backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain_openai
gitpython
docker
python-dotenv
httpx
pytest
httpx
pytest
